refactor(runner): extract dependency update flow
